+++ prisma/migrations/20240808210239_add_column_function_url_openaibot_table/migration.sql
@@ -1,0 +1,7 @@
+-- AlterTable
+ALTER TABLE "Instance" ADD COLUMN     "disconnectionAt" TIMESTAMP,
+ADD COLUMN     "disconnectionObject" JSONB,
+ADD COLUMN     "disconnectionReasonCode" INTEGER;
+
+-- AlterTable
+ALTER TABLE "OpenaiBot" ADD COLUMN     "functionUrl" VARCHAR(500);